Check for focus change on global mousedown
If the document is blurred (but still has a selection) it is
possible to clear the selection by clicking elsewhere without
triggering a focus or blur event, so listen to mousedown globally.

As we always check for a valid selection to trigger documentFocus/Blur
events it doesn't matter than many of these mousedowns aren't required.

Change-Id: Ic44e0f8efdca72a0ba4c5140d26cac909acaecaf