Apply 'default' or 'custom' size on load
When loading the scalable object into the MediaSizeWidget we should
make sure that the process updates according to either custom size
or default size.

There was a bug where the size widget showed the dimensions of the
image with 'default' button clicked instead of making the dimension
inputs empty in the default case. This code fixes the issue.

Change-Id: I01f4d1a4d2eaa3f1685d4dcd525372b869076b5c