Use this.$() in LanguageResultWidget#highlightQuery
It might be more characters than $(), but it makes the code almost
thirty times faster: on my machine, running updateLabel() for a full
result (e.g on init) went from 1770ms to 63ms. This appears to be
because jQuery/Sizzle ends up going through some really slow code
paths when we append nodes from one document to a parent from
another document.

Made highlightQuery() an instance method instead of a static method
so this.$() is available. It wasn't overridden anywhere anyway.

Bug: 65512
Change-Id: Ib33249f3c2598b75fa1cc21fb6aefaf5755564fe