Work around IE DOMParser bug in ve.createDocumentFromHtml
DOMParser in IE doesn't like empty strings. If you feed it an
empty string, you get a document whose .documentElement, .head
and .body properties are all null. To work around this, map
the empty string to '<body></body>'.

This fixes lots of unit tests in IE, because
ve.createDocumentFromHtml( '' ) is used a lot in the test suite.

This bug affects both IE10 and IE11 (older versions of IE don't
have DOMParser).

Change-Id: I78a00df2823d3b1c386b8356771708dadb7f22e8