Always refocus the paste target when a node is selected
The previous assumption was that if the paste target had a
selection it didn't need to be re-selected and focussed, however
it is possible for the paste target to keep a selection but lose
focus. As this event usually only fires when the paste target
loses focus, just remove the optimisation all together.

onFocusChange will ensure that no extra focus/blur events are emitted
(the reasoning given for this change in I466a64364).

Bug: 67178
Change-Id: I7c87b6da3aedf2923f75dc6f91c0d43b23847334