In the toolbar, display superscript tool above subscript tool
This ordering makes more sense to me intuitively. A quick glance at
other text editors also shows that, when they are placing these
options 'vertically', superscript always goes first.

It felt especially jarring in Polish, where the names for superscript
and subscript are essentially 'up-script' and 'down-script', and the
up and down were reversed.

(The change in ve.ui.AnnotationTool.js is incredibly stupid, but as
far as I can tell this is the only way. The toolbar doesn't permit
explicit ordering of the tools, only 'promoting' and 'demoting'
selected ones.)

Change-Id: I0e7c57328dcebf52876cbefa484f7d967fc63e43