Add toolbar menu labels to each group for consistency
As the mobile side of VisualEditor adds more features, we will want to expose an increasing
number of these standard function menus to the user. Additionally, a couple of the existing
groups (history and link) provide no current functionality for their labels to be shown, so
are included only for completeness (new labels for the history, paragraph formatting, block
structure and text styling groups; the 'insert' group retains its use of the existing label
in advance of a better solution being found.

To see these in action, visit the mobile demo (…/demos/ve/mobile.html#!/src/pages/h1.html),
and click on each of the toolbar groups to witness the majesty that is a semi-unstyled dump
which we need to improve. :-)

This also fixes the label for the text styling block in the commands help dialog to show up
as such, and not (confusingly) as "formatting", as a clash with the block immediately below
it is unfortunate and inconsistent.

Change-Id: Iec7b82b3e1382f269bf4c26f96d17c1d2ba1cb00