Fix calculation of affectedRange in transaction processor
Any insertion with two close nodes at the root (e.g. copy & paste
more than one paragraph) results in the whole document being
marked for rebuild. On large documents this is ridiculously slow.

To fix, decrement the minInsertLevel whenever we move into the parent
scope so we don't keep ascending the tree for siblings.

The fact 'minInsertLevel' exists as a variable but is never changed
suggests this is just a typo that has been lying around for 3 years.

Bug: 68301
Change-Id: Ifdcb29196e328e78a8273b0a385d3a3dff7d9144