Followup b141a7d23: don't display:none; the popup in the DesktopContext
We have to be very careful to hide the context and its popup using
visibility: hidden; rather than display: none; , because the latter
breaks all sorts of iframe-related things in Firefox.

This logic was mostly there already, but was broken by the great and
terrible dialog refactor (when DesktopContext's show() and hide() were
consolidated into toggle()). It's a mystery to me how the old code
even worked, because it looked like it should have had the same problem.

Bug: 68223
Change-Id: I54fbf03e91c2806548f6bc7fc9d9cfed53e6c90d