Highlight and bind events to $focusable, not $element
Still use this.$element for setting classes and properties.

This reveals a masked bug in nodes which override $focusable/$resizable
and use generatedContent, as GC re-defines this.$element but doesn't update
this.$focusable/this.$resizable.

To fix this add in getters for focusable/resizable to GC node, and use them
to redefine those elements on render.

Change-Id: I2415bbcdf94bbb25773a37e4f3a5048a567bb025