ve.ui.AnnotationInspector: Allow content insertion without annotating
Needed by Ie4a633ed in mediawiki/extensions/VisualEditor.

I think this makes a bit more sense. Previously, you could only insert
content if you were going to annotate it – now, you can always insert
it *unless* you are going to de-annotate it. (Without this condition,
clicking the 'remove' button of an inspector for a collapsed range
would insert content, which is silly.)

This is technically a backwards-incompatible change, but it doesn't
break anything in core or MW as far as I can see.

Change-Id: Icc2e368ac2f7951f53981d834029ef8b8f075d61