Work around getSystemPlatform() race condition
We have ve.ui.triggerRegistry.register() calls running from the
main file scope, which is evil for many reasons, but one of them
is that ve.init.platform may not have been constructed by that
time, causing JS errors because register() needs to call
ve.init.platform.getSystemPlatform() .

To hack around this, I made getSystemPlatform() and isInternetExplorer()
static, and wrapped the former with ve.getSystemPlatform() which
uses ve.init.platform if constructed and falls back to the
base class otherwise. A proper solution would involve not
running all these .register() calls in file scope, and cleaning
up the mess that is the platform system in general.

Change-Id: If014c52d7c4c4a1d2949ff262c35d64329d4f803