Followup b141a7d2: make inspectors not close immediately in IE
The great and terrible dialog refactor removed the 200ms timeout
at the start of the ready process for inspectors. This was a good
change, but it did break inspectors in IE.

Internet Explorer does not handle multiple selections in different
frames correctly. We wrap inspectors in iframes so we can have a
selection in the main document and another in the iframe, but
this doesn't work well in IE. Once an input in the iframe is focused,
the selection in the main document is invisible and inaccessible.

When the dialog is set up, content <div> is focused (the refactor
made the <body> be focused instead, but I had to change that back).
This causes a focusout event (and also a focusin event on the iframe
which shouldn't be emitted, but that's harmless in this case), which
gets debounced. At that point, asking the main document what
is selected will still return the correct result. Then later,
the input inside the inspector is focused, and from that point
on asking the main document what is selected will return null.

If there is no timeout at the start of the readyProcess, then
the debounced focusChange handler won't run until after the
rangy selection has been nulled. But with a timeout, it'll
run after the focusout event has been emitted but before the
rangy selection is nulled, which is exactly what we want
(and also means the code works completely by accident).
In the long term, we should look into making inspectors
resistent to this kind of behavior (and, by extension,
non-isolation, perhaps?), but for now readding the timeout
with a length of zero appears to fix this.

Change-Id: Ib8b6abf8a9b7b73105e68d02e122ed8afda68405