Followup 355be3ed: fix getOffset() behavior inside block aliens
If you gave getOffset() a DOM offset inside a block alien, it
would correctly return the DM offset inside of that block alien.
This was broken in 355be3ed which caused getOffset() to count
text node characters inside of block aliens, which led to some
very strange results.

Fix this by making the lengthSum addition conditional on
canContainContent again, as it was before 355be3ed. Kept the
isWrapped addition conditional on !isContent, that part of
355be3ed was correct.

This was causing bugs in cases where the user was able to get
the cursor inside of a block transclusion, which IE allows.

Change-Id: I5cfb1df45ed9ff1497cb179b0f5c2f442cf1e106