Fix unbalanced transaction when removing to/from the edge of an inline node
We (somewhat understandably) assumed that when canBeMergedWith() said two
nodes could be merged, we could perform a simple merge. However, there's
a quirk in canBeMergedWith() where m.canBeMergedWith( n ) returns true
if n is a content node whose content container parent is a sibling of m
(or vice versa). In that case, m and n aren't siblings and performing
a simple merge deletion (by deleting to the end of n on the inside)
will fail. Instead, we need to delete the entire content node
in that case (by deleting to the end of n on the outside).

Bug: T72375
Change-Id: Iea6c70e1b070c9de1d5cbb6754c24addabb38317