build: Make headers atop dist files suitable for distribution
Previously we prependied the same header in dist as for in-repo files,
which isn't very useful once distributed as a standalone file. Also,
change all in-repo files' headers to simply be the copyright message
with a reference to http://ve.mit-license.org.

Also fixed grunt-build which was broken. Grunt "git-build" should be
part of `grunt test` not `grunt build` – this way you run grunt-build
for releases and it won't include the git commit hash.
This matches the setup we have in OOjs.

Change-Id: I99acd6c6a6748b2029bb653836fed00dd8b5d21b