Emit change event when dimensions change in MediaSizeWidget
Whether the dimensions are valid or not, the 'change' event should
be emitted when the dimension inputs are changed. This will cause
the dialog to revalidate the size widget in general and hide or
show the 'apply' button accordingly.

All changes for the widget value change the inputs themselves,
either by clearing them out or setting a different value. The
'hidden' changes (like those that set the max dimensions,
default dimensions, etc) are changes to the Scalable object
that emits its own changes.

Bug: 70861
Change-Id: Idb82d597d5b2b6664401208a54d33d26edbf6d31