Followup 2432cbf: add recursion guard to updateSlug
popStaging() changes the model selection, then onModelSelect()
calls updateSlug() which calls applyStaging(). This is bad, and
the only reason this doesn't fail spectacularly is because we
didn't have multiple active staging levels.

The only reason we even noticed this is because we were
re-getting the range from this.slugFragment, but this.slugFragment
had been nulled out by the recursive call to updateSlug(),
so we got a JS error. This fixes that, but we also
already had this range lying around, so use it.

Bug: 72095
Change-Id: Ie50805a12ec2af3f0e1d97a4e64751a9b74fc422