Use local var instead of bind() for inline functions
Follows-up d3f26e653a4c. Using .bind() makes sense for binding a
function reference, but when the function is inline, might as
well use a local variable.

* Reduces code complexity.
* Makes code slightly faster.
* Easier to read (no need to visually track which 'this' this is,
  variables have descriptive names).
* Easier to maintain.
* Better static analysis.

In terms of memory holding, the result is the same since inline
functions naturally hold on to the containing closure due to
lexical scope.

Replaced all instances of "}.bind" in src/.

Left instances of "identifier.bind" alone.

Example:

	var widget = this;
	promise
		.always( function ( opened ) {
			opened.always( function ( closed ) {
				closed.always( function () {
					widget.foo();
				} );
			} );
		} );

Instead of:

	promise
		.always( function ( opened ) {
			opened.always( function ( closed ) {
				closed.always( function () {
					this.foo();
				}.bind( this ) );
			}.bind( this ) );
		}.bind( this ) );

Change-Id: I797a3667c67d52568150be9be5d043d149f22077