Don't create tables with a thead
There is no way for the user to see the difference between
thead and tbody, no way for them to manipulate this difference
(whether a cell is header/data is on a cell level), and
Parsoid doesn't understand <thead>s. Eventually we should
deal with this correctly, but right now that isn't feasible.

Change-Id: I7a45d42f761ab982f59a939b6368c5e59e4dea7a