Fix selection transitions between FocusableNodes and non-linear selections
Selection transitions from NullSelection to FocusableNodes
were broken: if the selection was a NullSelection and you
clicked a FocusableNode, nothing happened.

Selection transitions from FocusableNodes to TableSelections
were also broken: if you had a FocusableNode selected,
then clicked into a table, the FocusableNode would continue
to be highlighted as if it were selected.

Bug: 72721
Change-Id: I862da3eb304edf50b82c16c7b0ab735f74d003bd