Handle empty rowspan/colspan attributes correctly
This commit Should(TM) fix all remaining rowspan/colspan
corruption issues.

No longer store rowspan/colspan values of NaN. Instead,
omit the rowspan/colspan attribute when the attribute
value does not parse to a number or is empty
(annoyingly, '' parses to 0). Always set the originalRowspan/Colspan
attribute if an attribute value was set, and use it if
the original value parses to the same number as the
rowspan/colspan (as before), or if rowspan/colspan
is not set (replacing the NaN check we used before).

Change-Id: I5db816071dccb7eff015509c7ce01b80758a47db