Fix whitespace consistency checking when child has no whitespace property
Our whitespace consistency checks never check innerPre, because
we assume it was already checked when the first child's outerPre
was processed, and then code then clears innerPre. So if innerPre survives,
we assume that must be because there was no children. This is almost
true, but if the first child didn't have any whitespace data at all,
the code path that cleared it was never taken. In this case, make
sure to clear it as well.

EXCEPT if we're doing this from a content node, those never have
whitespace recorded on them. Of course. Sigh.

Change-Id: I8ad147f174040b7f9223c57ac8583c22e172a3d7