ve.ce: Remove hacks from Surface#onDocumentKeyPress; add related IME test
We had a hack in place to filter out non-character keys, to prevent
unexpected content deletion (caused by pawning) when selection is not
collapsed and the user presses, for example, the Home key (Firefox
fires 'keypress' for it). Pawning is gone since 40101e42 and this
seems no longer needed.

Added an IME test for the described situation.

Bug: T69367
Change-Id: I277b92b9d11bfda01a92f3e615ca674feb920a1a