Replace ve.dm.Model#matchesAttributeSpec
Attribute matcher specifications can take any form from the list below:

* Object with an array 'whitelist' property, containing any mix of booleans,
  regular expressions, and strings.
* Object with an array 'blacklist' property (same composition as above).
* Object with both 'whitelist' and 'blacklist' properties (ditto).
* Array (ditto).
* Regular expression.
* Boolean.

This flexibility is purchased at the cost of a lot of branching in a very hot
code path (ve.dm.Model#matchesAttributeSpec gets called for every property of
every element).

This patch allows attribute specifications to be either a boolean value or a
function which takes an attribute name and returns a boolean value. This allows
for dynamic and complex specifications without incurring costs for the common
cases.

Change-Id: Ia0e0fe4fc2385dab67c816148cfc362a39413dac