Guess annotations better for complex changes
For complex changes with no deletion, use the annotations from the document
offset before the start of the change (which means none if the change is at
the start of a CBN). Previously, the annotations were always taken from the
start of the deletion, which would always give none if the deletion was empty.

Also document the change detection heuristics better.

Change-Id: I0fccff4c8bfd2b74fa1fd16a205cd44800478338