Make annotation inspector getMode method trim
In cases where you selected a linked word plus adjacent spaces, the
inspector would be in insert mode instead of edit mode, because
getSelectedModels() only returns models that cover the whole selection
(unless you use the all argument, but we weren't doing so here because
that's farther from the behavior we want). This differs from how
annotation inspectors actually work because they have special behavior
in getSetupProcess that trims the fragment before deciding if there's a
model to be edited. By adding trimming to the getMode() method we can
ensure that insert mode is used correctly.

Change-Id: I354059843b770fb12996bf9473c29daa6952de7e