When replacing e.g. heading1 with heading2, use attribute changes
Previously we would destroy the entire heading1 node and create a
brand new heading2 node, without copying over attributes or
originalDomElements or anything. This meant that if you
had <h1 style="background: red;"> and changed it to an h2,
you'd get a plain <h2> with no red background.

I didn't have to add any code to deal with changes to the
level attribute on headings, because all that infrastructure
was already in place, it just wasn't being used.

Change-Id: I22fd03b8b2eab5eebf735b7097a956757f2aac19