Make transaction processing exception-safe
Make modifier methods return a function that reverses
the changes they made. Then if an exception occurs during
or after applyModifications(), we can just call these
functions in reverse order.

If an exception occurs during synchronization, then the
tree could be in any sort of crazy broken state, so
we rebuild it from scratch.

In the future, we should really dispense with the distinction
between operations and modifications. The only reason we need
a modifications layer right now is because operations as currently
implemented are not safely reversible: trying to reverse an
operation would also cause synchronizer actions to be queued,
and in order to compute these actions, queries are made to a
half-baked tree which will probably cause exceptions.

In the tests, assert that the linear model and the tree
are unchanged after an exception.

Bug: T70892
Change-Id: Ic7d0088861cac9d1873f66633402e1b7eef2d645