ce.RangeState: Check this.node for falseyness, not just null
I don't know why and I don't have steps to reproduce, but I managed
to get into a situation where this.node was undefined rather than null.
That's probably a bug in itself, but as long as that can happen we
shouldn't assume that this.node !== null means it's safe to access
properties of this.node.

Change-Id: I9490077074e4249706459c92c16b55cf8c086836