FragmentInspector: Execute action on enter, rather than closing with data
Closing with data is evil and should die in a fire.
As long as it hasn't yet died in a fire, the 'done' action in
FragmentInspector is already wired up to do
this.close( { action: 'done' } );, so code using the close-with-data
pattern will keep working. However, subclasses can now override the
'done' action and possibly decide not to close the inspector when
enter is pressed.

Change-Id: I523d8e19eeabee0e2c88870d5da448d5442dee0e