Don't drop whitespace when removing empty slug paragraphs
If you had two block templates with whitespace between them,
then used the slug between them to insert a paragraph, but
left that paragraph empty, the paragraph wouldn't be output
(because the converter removes empty slug paragraphs) but the
whitespace would be dropped.

This happened because we didn't work hard enough to pretend
that removed slug paragraphs never existed: they were still
considered for the purpose of verifying whitespace consistency.

Bug: T93129
Change-Id: I14f80ed4ccec21f43acf4f4e5078349f0742fe95