Split handlesOwnChildren and ignoreChildren
handlesOwnChildren did two things: it caused the converter to
allow the node's toData/toDom functions to handle the node's
entire subtree, and it caused the node's children to be skipped
by various pieces of code.

This commit splits this so that handlesOwnChildren is only used
for conversion purposes (and for attribute rendering), and
the new ignoreChildren property is used to signal that the
node shouldn't be descended into.

Existing nodes with handlesOwnChildren were migrated to have
both handlesOwnChildren and ignoreChildren, so their behavior
doesn't change. The only exception is InternalItemNode, which
doesn't need handlesOwnChildren because it doesn't interact
with the converter.

Change-Id: I800459aacd7d31a77f86f280d82a9ffe701c8d31