Fix RangeState detection of selection outisde of document
Previous logic was mistakenly treating 'selection' as if it
was a real native selection, which it isn't. Fix logic to use
the correct variable, and detect if you are inside the document
node the same way we do in ve.ce.Surface#onFocusChange.

Refactor RangeState slightly to pass in the documentNode only.

Change-Id: I37903f4f79fa77d1cf1def6b30f90c0592675c74