Make DM nodes sensibly hashable
Currently we only hash annotations (for comparison) and aliens
(for store indexing). If we were to hash other models
(i.e. nodes for serialization) we would not want to full inner HTML,
e.g. <p>Foo</p>, but just the root <p></p>.

The exception is AlienNodes where originalDomElements contains
the information required to rebuild the DM HTML.

Change-Id: Ie916b88e6032dff04d7bb1831a0800033c80e5ab