Make triple-click expand to CBNs, not BranchNodes
We used to expand to the nearest BranchNode, but really this
only makes sense for CBNs. Restricting ourselves to CBNs also
ensures that adjusting by (1, -1) is sane (for the DocumentNode
it isn't).

To deal with selections like |</p><p>|, collapse the selection
before expanding. This doesn't change the end result for selections
entirely contained within one CBN, but it helps us not go crazy
when the selection spans a CBN boundary.

Bug: T93461
Change-Id: Ib7338c6493963ae1f7f092739ac4ec90107a0ae1