Replace abstract functions that just throw with null
Per Timo, this avoids us having pointless code that is never
executed, and the replacement notification to the user (that
ve.blah.UnicornAnnotation.foo is null and not a method) is
sufficiently clear that the issue lies in the lack of existence
of the ve.blah.UnicornAnnotation#foo method.

This removes 46 such methods from our codebase (1970 -> 1924),
and so incidentally boosts nominal code coverage for the functions
from 60.61% to 62.06%. Progress.

Bug: T54482
Change-Id: I9480e7375cb3e3487fde19634b1f545131ee5dde