Add a default 'cancel' action to FragmentInspectors
This will automatically add a default cancel action to the common
modes in the fragment inspector. However, we have to override this
functionality in the annotation inspectors, since those don't really
cancel their operation on dismissal.

If a new inspector uses any other arbitrary modes and overrides the
action set given, that inspector should be responsible to include
the safe 'cancel' action for usability.

Bug: T96303
Change-Id: Id9e9c77c8a1d7cad582c1a6c0408390e08008a1c