ve.utils: Add ve.fixBase() utility
Utility to convert relative base URLs to absolute ones.
Based on code in VE-MW, but changed significantly.

In particular, this one has a fallbackBase parameter which
fills in a base URL if the document doesn't have a <base> tag.
We have to use the presence of a <base> tag rather than .baseURI
because in Firefox (but not in Chrome), secondary documents inherit
the base URI of the main document.

Change-Id: I607fa5797334dda6db11c6c0799d2c9e5024d993