Support RegExp sequences; trigger sequence matcher after newline
Allow a ve.ui.Sequence to be specified using a RegExp, but hard-limit the
size of the string checked by the RegExp to avoid O(N^2) blowup when
inserting N characters of text.

Tweak the signatures of the sequence match methods to return the range of
the matching portion of the data, to avoid having to repeat the text
extraction and regexp match when executing the associated command.
Allow passing this range as the current selection when executing the
sequence's associated command.

Trigger `checkSequences` in `handleLinearEnter` so that sequences
are checked when enter is pressed (pressing enter doesn't trigger
onSurfaceObserverContentChange).

Change-Id: I4d03d2f7b437ea4a289a4c5cee2a081998228319