Make focusing surfaces work in Firefox
When you called .focus() on a blurred surface, the toolbar
would be enabled, but the cursor wouldn't actually go into
the surface, and keystrokes wouldn't be picked up.

This happened because the .focus() code path for NullSelection
relied on showSelectionState() to do a DOM .focus() call.
However, showSelectionState()'s check for whether it should focus
a DOM node was broken: when nothing is selected,
document.activeElement === document.body, so
OO.ui.contains( document.activeElement, newSel.focusNode ) is true
(because everything is inside the <body>), so a DOM focus
never happened.

Instead of checking whether newSel.focusNode is inside the currently
focused element (which has all sorts of other problems aside from
the default being the <body>), determine which node we would focus
if we needed to, then check if activeElement is inside it.

Additionally, make .focus() not rely on showSelectionState()'s focusing
behavior, but instead have it normalize NullSelections to the first
content offset, and then proceed to focus the surface. This will
cause a double focus so long as showSelectionState() behaves the
way it currently does, and it'll invoke onModelSelect() twice,
but that shouldn't cause any problems. In particular, showSelectionState()
short-circuits when the selection is already in the desired state.

Bug: T109805
Change-Id: I33fefbccd5c4e5abae77b01d14238257e1360799