Make fixBase tests account for no-op behavior when baseURI is valid
Chrome does not consider a protocol-relative base URI to be valid
for a synthetic document, but Firefox does (because FF considers
window.document to be the parent, but Chrome considers such
documents to be parentless). For this reason, the first test case
behaves differently in FF and in Chrome, but that means fixBase()
is working by design: it's not supposed to mess with already-valid
base URIs. Adjust the tests to expect this behavior.

Bug: T113595
Change-Id: I6adddf80feda262a71156a13cc23ea7b363f365a