[BREAKING CHANGE] GeneratedContentNode: Introduce new hash for rendering
Generated content is stored in the IndexValueStore using
the model hash, however sometimes changes in the model
hash don't produce a change in the generated contents so
we can improve this in some cases by allowing nodes to
simplify the hash they provide for this.

e.g. A MW extension that provides an attribute to set a
link ID can exclude this from the hash as changing it
results in an identical rendering.

Keep this separate from the regular getHashObject in case
we want to have a hash the describes the state of the model
for some other purpose.

Bug: T114689
Change-Id: I5d42943a06f63acab828ccad39802df573da064a