Pass staging state through Document#commit to ve.dm.Node update event
This allows GeneratedContentNodes to know if the update was a
staged update, and choose whether to render the error in the view.

Staged updates only cause the node to change the colour of the focus
highlight, whereas full updates will show the full error in the view.

This makes generatedContentsFirstRender redundant as we call
update with staged=false on initialize.

Bug: T114480
Change-Id: I8ebb26d4bcd69e611c5bc2527efb335098938172