Follow-up I553e31: Don't show error for non-rendered GCNs
To avoid all issues with generatedContextValid meaning 'true'
when it is undefined or null, invert it to '..Invalid' which is
only true when we know we have a GCN which has rendered and failed
validation.

Also always perform this check, even on first render as the node
may load in an error state.

Bug: T114527
Change-Id: Ia08d8e7b74e37edc5e73b2ed8adc8ece90764ada