getViewportDimensions: Fix calculation of 'bottom'
When the top of the surface is in view, 'top' is clipped
to 0, so we can't calculate 'bottom' from this value.

Instead compute it as the height of the viewport added to
the surface scroll offset (-rect.top).

This fixes issues with DesktopContext floating near the
top of the page.

Change-Id: Ib0c06fe0d82eda1abb9151356025de995e5c0f45