[BREAKING CHANGE] Refine VisualEditorSupportCheck call and document pattern
Follows-up I27a8e7b437664.

Call the global as-is instead of via 'window'. This isn't a Window class method
and should not be called as such. It's a fully-fledged global function can be
called as-is. This also benefits JSHint detection. It not being whitelisted, but
used via 'window' made it more error-prone (e.g. when mistyped). Now it can
detect it properly.

It was also wrongly indexed as a class when it is in fact a method.

One could create a private mixin for global to allow multiple global methods
to be documented in categories, but for now indexing 'global' as a single group
in the categories file under Initialisation.

Now that the warning for having a global method is triggered, surpress it with
"-global" in the jsduck.json file.

Change-Id: Id53bdfe1c459ea5995e409644d1b7afddf11227c