ve.test.utils: Don't require QUnit expect() anymore
This hasn't been useful in QUnit for a over a year now with the improved
assertion context object and tracking of asynchronous tests without
shared global state.

It's also causing a conflict because VisualEditor-MediaWiki includes
this file and ends up overriding MediaWiki core's loser configuration.
And while MediaWiki still happens to pass requireExpects in REL1_26,
Wikibase has started to write tests that don't set expect() for each
test (which is fine).

Change-Id: I4929c10424be7bb8f717f8b01f1c139a25ad117f