Transactions: Ignore internalList changes in getModifiedRange
getModifiedRange is used to calculate selection translations
so we don't care about the internal list which isn't rendered.

With a newFromDocumentInsertion transaction you may get an insertion
in the main document and the internalList and in these cases
we still want to know the offset translation in the real document.

Bug: T129539
Change-Id: I3cac35a4fa4a837b15e654e59e0eac680d69391e