Table editing: Refactor active table node to be more generic
The functionality the active table node provides (limit selection
to a certain node) will be useful in other situtaions, so make
it more generic.

Makes the cell node the active node, instead of the table node, so
the available range can be computed directly. When the active table
node is required, just use tree traversal.

Change-Id: Ifd1970abb12b6597284b06e0090591668ee73c3c