ve.ce.FocusableNode: Wrap updateInvisibleIcon in rAF
The width/height measurements done by hasRendering() cause slow
forced reflows. Deferring these calls causes 40% less time to be
spent in them, but we still spend ~400ms (down from ~700ms)
doing reflows caused by these width/height measurements.

I think this is probably a layout thrashing thing, as we modify
the DOM (add the icon) for each FocusableNode that turns out
to be empty, so that probably invalidates the layout and
causes a new reflow to be needed for the next node.

Change-Id: I29d64b1940eb8ca9b7e4825c68427f3dbd8d87f5