TransactionProcessor: Fix bug where annotation events could be emitted on the wrong nodes
DocumentSynchronizer#pushAnnotations takes an unadjusted range, and is used
that way in the DocumentSynchronizer tests, but TransactionProcessor gave it
an adjusted range. This caused annotation events to be emitted on the wrong
nodes if a replacement and an annotation occurred in the same transaction.

These kinds of transactions never happen in practice, but it was difficult
to figure out whether TransactionProcessor's usage of pushAnnotation() was
correct until I added tests that proved it was incorrect, so we might
as well keep those tests :)

Change-Id: Ie4888c76a3de1143cfb30ce2f7cf3ab1af16f5ba