ve.ce.Surface: Prevent horizontal scrollbars when hovering FocusableNodes in RTL
.ve-ce-focusableNode-highlight-selectable was making the document 3px wider,
because it contained an nbsp and was positioned in the top left corner.
This was because we neglected to put both left:0 and right:0 on
the .ve-ce-surface-highlights container. However, that container
doesn't need any positioning rules at all (because it's inside
localOverlay which is already absolutely positioned), so just remove them.

Change-Id: Ib23c56bbd072979fdebb80a6aa09df394e7ab06b