DesktopContext: Unhide context before positioning popup
The popup is inside the context, so if the context is hidden
the popup is also not visible, which breaks its positioning code
in center mode (because measuring its width returns 0).
Ensure that the popup is visible while it's being positioned.

In combination with I37f06ae7b in oojs-ui this fixes a bug
where contexts for right(/left)-aligned focusable nodes expand
the document and cause horizontal scrollbars.

Bug: T159472
Change-Id: I9b5107682b2ec88d0a66aee1fb86507e587ec3e6