ve.ui.TableAction: Don't pre-build sequences of transactions
Previously, arrays of txs bere being built with TransactionBuilder, before
applying the whole array in sequence. This was giving invalid final retain lengths,
because each tx should be built only when the document is in the correct start state
for applying it.

This change instead creates arrays of txBuilders - zero-argument functions that
return a transaction fitting the current document state. The output of each txBuilder
is then applied to the document before calling the next txBuilder.

Change-Id: Ic9db29b982585a75efdc3baaaa642a636fc5cd44