VisualDiff: Fix doc child replacement
If types are not equal, mark the change as a replacement, instead
of just setting typeChange, which is never used.

Only use 'structural-(insert/remove)' if a non content node is also
a branch node (so leaf nodes like AlienBlockNode aren't structural).

Change highlightRemove/highlightInsert order to ensure the node
at the correct offset is annotated.

Bug: T166801
Change-Id: I237254096ef58fca6060ef45a29b34244410aca5