Fix infinite loop in ve.BranchNode#getNodeFromOffset
This is a minor clean up of the refactoring done in
1cbcc2f0bf66fd824c255a7e6ca1e87fdacd46e9.

If the document is empty (consisting only of ve.ce.InternalListNode)
then it is possible for `currentNode.children.length > 0` and
`currentNode.children[ 0 ] instanceof ve.ce.InternalListNode`.
This sets up an infinite loop, unless `offset === nodeOffset`.

This was being triggered when an empty document consisting only of
`<paragraph>, </paragraph>, <internalList>, </internalList>` was
p-unwrapped; for example, prior to paste in
ve.ui.MWWikitextStringTransferHandler.js or prior to update of a
`-{}-` construct by ve.ui.MWLanguageVariantInspector.

Change-Id: I933bf28637f3b3c235aa4b9f8d96d1d09a7b3085