Unbreak tests after Chrome upgrade
Chrome appears to have changed how many newlines it inserts when copying
text at an <h1>/<table> boundary. This change makes the test pass in
Chrome again but breaks it in Firefox. However, a few of our other tests
break in Firefox already, so I guess that's fine for now *sigh*.

Change-Id: I2fb767a80ebad7a2839b2b689b1035ccde41ecf7