rebaser: Fix sequencing of server operations
pendingForDoc was not actually ever being populated.
Fixing this sequencing resulted in the artificial delay (typically 5s)
being applied to every event consecutively, which is not what we want
(it's *very* slow), so make the delays run in parallel with waiting
for the previous step (and its delay) to finish.

Change-Id: Ia7a53600d2b0626f81f306aeb72cecbbe00fe7ac