Do less work when building the diff document
Since DiffElement works with copies of the old and new documents,
we can shallow clone these each time we make a slice.

Also add the attributes to the outer nodes within the converter,
since this is quicker.

And where a whole child of the document is inserted or deleted,
add in the <del> and <ins> tags directly, rather than annotating
every single character.

Bug: T187070
Change-Id: I3cce09054d77918ee79592baeedb29fae6325e46