ve.Node: Make #getOffsetPath not fail while rebuilding the document
If this DM node's document is in the process of being rebuilt
(ve.dm.Document#buildNodeTree), this.doc will still refer to that
document, but the node will be attached to a different, temporary
one. (I don't know why this is the case, but it's true... it is
very confusing.)

This caused #getOffsetPath to decide the node was not attached,
causing exceptions in ve.dm.MetaList#onNodeAttached, which
naturally expected the just-attached node to actually be attached.

Make #getOffsetPath look for any DocumentNode (consistently
with the function documentation), rather than the specific one
accessible via this.doc.documentNode.

The node offset paths from two different documents may not make
much sense for the ordering in ve.dm.MetaList#onNodeAttached, but
this should be okay, as the old nodes are going to be removed when
we're done creating the new ones.

Bug: T187690
Change-Id: I95cdd53bd29fd4686ff9012a6e673bee1510c0db