ce.Surface: afterPaste should always remove ve-attributes
If afterPasteAddToFragmentFromExternal detects that we're pasting from another
VE instance and an important element has been in some way mangled by regular
paste flow, it throws away all the pasteTarget sanitization and just accepts
the clipboard HTML as-given.

In this case, we should still clean up data-ve-attributes, as it's some of the
copy/paste scaffolding added by onCopy.

Bug: T196254
Change-Id: Ibbd31a784ca9d57e4c5cc7b75267b823cb56a044