TabIndexScope: ignore aria-hidden outside the tab scope root
Non-modal WindowManagers are currently always flagged as aria-hidden, so for
dialogs like the find and replace toolbar, this couldn't find any elements
that were tabbable ever.

Regardless, it seems to make sense to say that if the focus is already within
an aria-hidden element, we might as well treat it as being visible.

Bug: T197250
Change-Id: I0eed6c17385ed052cb220def7772e345461fe1e9