ce.FocusableNode: suppress left clicks within the focusable element
Quirks of the highlight positioning mean that it's occasionally possible to
click under the highlight. If there's a link clicked, page navigation would
occur, as the only handlers on $element or $focusable are (a) mousedown, and
so don't suppress a later click event, and (b) in the common case where
$focusable == $element, never call preventDefault.

As such: always suppress left clicks within $element.

Bug: T197659
Change-Id: I015fcd415e25e064528d76fab4840f003843566c