ve.ce.Surface: Handle null native selections in #getSelectionState and #showSelectionState
Turns out that it is possible to have a degenerate document which has
no cursorable offsets (you can't put a text cursor anywhere, the
native selection is null). ContentTranslation documents are like this
(until you start translating them, which inserts editable text). They
consist only of leaf nodes that can't have slugs before or after.

You can still set selections on it programatically, and when
VisualEditor attempted to render them as native selections, things
broke. Since ContentTranslation depends on it, we have to be able to
handle this situation. We don't have to handle it well, but we can't
throw exceptions.

After this patch, #getSelectionState can return a null selection state
(instead of failing), and #showSelectionState will clear the native
selection when given a null selection state (instead of failing).

Added a testing-only node that acts like CXPlaceholderNode, and test
cases using it that show these functions don't fail.

Bug: T199011
Bug: T199754
Change-Id: I97eb7997bb22456a939dcb3894675e8b7ebe9c19