ve.dm.Document: Replace empty nodes after inserting non-text data
If trying to insert non-text data (e.g. `<horizontalRule></horizontalRule>`)
inside a node which must contain only text (e.g. `<paragraph></paragraph>`),
we split the parent node (and any of its ancestors, as needed) until
we're able to insert the new data.

There is a special case: if the insertion position is the first/last
position inside the parent node (so the inserted node would be its
first/last child), we just move the insertion position one step to the
left/right, to avoid creating an empty node by splitting the parent
here.

This patch adds a new special case: if the insertion position is the
first *and* last position inside the parent node (so, the parent node
is already empty), we will now replace it entirely (and not leave the
empty node).

Bug: T201573
Change-Id: Ic66d2f49d57c430674e33f0880cfa7775c2761d8