ve.dm.Document: Remove incorrect handling for combining characters in #fixupInsertion
As a result of this code, we put two Unicode codepoints into a single
model position. As a result of that, model offsets and CE DOM offsets
do not match, and all hell breaks loose.

This was added in 90e4f4d665a635e753d8fef160e93f75f5287668 and probably
should have been removed in 6f2090aac63b31661748ef0db3ff1edd330656ab.

Additionally, remove the code that tried to prevent unattached
combining marks at the beginning of a line. It only prevented
non-annotated combining marks that were the first character of an
insert operation, and it seems like a very confusing behavior anyway.

Bug: T85093
Change-Id: I6567b4290b0f9c850157a464d19bd6b8ad1982b8