ve.dm.ModelRegistry: Fix extra spaces in rel="" attributes preventing matching
Syntax like rel="foo  bar" (with two spaces) would be parsed as three
node types: 'foo', '', and 'bar'. The empty type would then cause
matching to fail if it was not listed in 'allowedRdfaTypes' for a
node.

Currently this is not an issue in practice, because there aren't any
nodes with additional 'allowedRdfaTypes' for which Parsoid would
produce HTML with extra spaces. But it can produce something like
below (for a template-generated ref), so this seems like it could
cause bugs in the future.

    <sup about="…" class="mw-ref" id="…" rel="dc:references"
      typeof="mw:Transclusion  mw:Extension/ref" data-mw='…'>

Change-Id: I1f467f51017e2deae30905163bf5e6b07048cecf