ve.ui.MediaSizeWidget: Fix conditions for when we need error messages
We may need to display an error even if we are not validating the
input against original dimensions, e.g. if the input is not numeric.
Move the code for error label outside of noOriginalDimensions check.

Also move the commented-out code for the scale/"upright" input outside
of noOriginalDimensions check. The scale in it would be relative to
the default thumbnail dimensions, rather than original dimensions.

Change-Id: Iebfbe0b06c0d7ba1c7f7e121117c8235d127eeba