ve.ui.DesktopContext: Actually hide mobile '.ve-ui-linearContextItem-info'
The selector '.ve-ui-contextItem-description' should have been
'.ve-ui-linearContextItem-description' (it was not updated in
fa80099dd22730d0c8293b09cd132a35bdca0f6c), but I think what we
really want is to hide '.ve-ui-linearContextItem-info' (which is the
parent node of description).

This element is always empty on desktop (on mobile, it is used
instead of '.ve-ui-linearContextItem-body'). Its existence makes it
more difficult to style the header of desktop context items.

Change-Id: I82c8ecdbedbd3997b0fd6c88cc0438d9a9ac9049