Remove -moz-user-select:none from hack and buttonElement
The hack addressed an selection issue in T70537, but selecting
across moz-user-select:none causes FF to create a multi-range
selection, which we really want to avoid supporting.

The original issue appears to not be reproduceable anymore, and
even if it was, it seems to be less severe than the problem it
causes, which prevents users from creating selections.

OOUI buttonElements also apply user-select:none, so override
this in FF when dragging a selection.

The * selector in the hack was also causing caption nodes
to be unselectable in FF as there was no exception for them
(that functionality was added a long time after the hack).

Bug: T214976
Bug: T217047
Change-Id: Ieef78dab626fa9f20fe188122929228acfad7820