Distinguish between $scrollContainer and $scrollListener
When scrolling the whole page, $scrollContainer should be
the <body> or <html> element. Currently it is `window` which
means $.animate doesn't work.

$scrollListener is the element which emits scroll events, which
in the case of the whole page is `window`. In all other cases it
can be the same as $scrollContainer.

Before this change
 `target.$scrollContainer.scrollTop( 100 );`
worked, but not
 `target.$scrollContainer.animate( { scrollTop: 100 } );`

Now they should both work.

Change-Id: I3ec11c384c590c47847cd5a4ad7369475a82a659