SurfaceSynchronizer: Force websocket transport
If we don't do this, the client will sometimes choose to use XHR, which
suffers from crashes when we try to transmit non-ASCII characters like
emoji. Also, paste buffers sometimes end up in the hash store for some
reason, and those contains U+2600 BLACK SUN WITH RAYS which also
triggers the encoding bug.

Change-Id: I07cee3c0209e48ea571f1451fc62302a6e3623ec