ve.ui.AnnotationInspector: Select inserted elements (not text) after inserting
When the inspector inserts text before annotating it, the cursor is
placed after the insertion, so that the user may easily keep typing.

But when the insertion is an element and not text, it's probably a
better idea to highlight it, so that the user can see what exactly
they just inserted. Currently the only case for this is inserting a
'link/mwNumberedExternal' in MW.

Also, remove a misleading no-op call to adjustLinearSelection(). The
method returns a new fragment rather than modifying it in place, and
we don't use the return value, so this did nothing. And if we did use
it, it would be incorrect, as the argument should not have been
negated.

Bug: T227630
Change-Id: I848aef8730f0b3c5bdec0ef87edf82f2665c7a9e