Fix ve.dm.Change#toJSON to cope with arguments passed by JSON.stringify
In 22eb0e4cac11243be2fbbf60ba2746eddbc733a7 , ve.dm.Change#stringify was
renamed to ve.dm.Change#toJSON . But when JSON.stringify calls .toJSON
it passes an index / property name as the first argument, which was
confusing the method. That was making stringifying fail if the change
contained an annotation.

Change-Id: I036f6ebce427e74a69ea6a9bf1470f8d32ff2a87