ve.ui.DesktopContext: Reserve space above anchored inspectors
Previously, when an inspector had the maximum allowed height, it
exactly filled up the space between the toolbar and the bottom of the
window. Scrolling the inspector into view would therefore scroll the
relevant selection out of view, and scrolling the selection into view
would scroll the inspector (partially) out of view. We apparently try
to do both, and thus constantly scrolled back and forth.

Bug: T229900
Change-Id: I984ae1840c706b8a6a3af200f56ff6fb2e9bdb28