Mark surface as "showAsDeactivated" when opening a window
Confusingly this doesn't change the visible behaviour in a regular
VE surface with nullSelectionOnBlur disabled. This is because without
'showAsDeactivated' set, the surface got immediately re-activated
but the first staging transaction of the window, then de-activated
again by the native blur event on the surface (which *did* pass
showAsDeactivated).

On surfaces that did set nullSelectionOnBlur this just caused them
to break (setting the model selection to null which stole selection
back from the inspector).

It is not ideal that a flag that is ostensibly for UI rendering
only (showAsDeactivated) changes the focus changing behaviour so
significantly so ideally this would be made more robust in the future.

Bug: T236400
Change-Id: I5fbf7c047a78bf16e955de5c1f6140a892a22f0e