Fix CodeMirror overlay alignment on new lines
On Firefox, when a user starts typing on a new line containing an inline slug
with a chimera, the chimera styles cause the CodeMirror overlay to be misaligned
relative to the VE source editor surface. This causes the text to appear blurred
on the affected line.

Adding a 2px negative margin to such chimeras in Firefox when CodeMirror is active
seems to successfully eliminate the misalignment and thus fix text display.

This is a port of If7e902d05e096f58ee8dc77dfe9a69bf3bd49882, which originally
aimed to apply the fix in CodeMirror's stylesheet.

Bug: T198278
Change-Id: Ice36dc27fc107322da47427596178c84d253c716